new logic for limit
