+++ src/programs/voice/voice-limit.ts
@@ -1,13 +1,12 @@
 import { Message } from "discord.js";
-
 import Tools from "../../common/tools";
 import {
   Command,
   CommandHandler,
   DiscordEvent,
 } from "../../event-distribution";
-
-const maxLimit = 10;
+import VoiceOnDemandTools from "./common";
+import { maxLimit } from "./common";
 
 @Command({
   event: DiscordEvent.MESSAGE,
@@ -20,13 +19,9 @@ const maxLimit = 10;
 })
 class HandleLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
   async handle(message: Message): Promise<void> {
-    const requestedLimit = Number(message.content.split(" ")[2]);
-    if (!requestedLimit) {
-      await Tools.handleUserError(message, "The limit has to be a number");
-      return;
-    }
+    const requestedLimit = message.content.split(" ")[2];
 
-    const channel = await Tools.getVoiceChannel(message.member);
+    const channel = await VoiceOnDemandTools.getVoiceChannel(message.member);
 
     if (!channel) {
       await Tools.handleUserError(
@@ -36,17 +31,14 @@ class HandleLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
       return;
     }
 
-    const limit = await Tools.handleLimitCommand(
+    const limit = await VoiceOnDemandTools.handleLimitCommand(
       message,
-      requestedLimit,
-      maxLimit
+      requestedLimit
     );
 
-    if (!limit) {
-      return;
-    }
+    if (!limit) return;
 
-    await Tools.updateLimit(channel, limit);
+    await VoiceOnDemandTools.updateLimit(channel, limit);
     await message.reply(
       `Successfully changed the limit of your room to ${limit}`
     );
@@ -64,7 +56,7 @@ class HandleLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
 })
 class HandleShrinkLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
   async handle(message: Message): Promise<void> {
-    const channel = await Tools.getVoiceChannel(message.member);
+    const channel = await VoiceOnDemandTools.getVoiceChannel(message.member);
 
     if (!channel) {
       await Tools.handleUserError(
@@ -75,7 +67,7 @@ class HandleShrinkLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
     }
 
     const limit = Math.max(2, channel.members.size);
-    await Tools.updateLimit(channel, limit);
+    await VoiceOnDemandTools.updateLimit(channel, limit);
     await message.reply(
       `Successfully changed the limit of your room to ${limit}`
     );
@@ -92,7 +84,7 @@ class HandleShrinkLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
 })
 class HandleUpLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
   async handle(message: Message): Promise<void> {
-    const channel = await Tools.getVoiceChannel(message.member);
+    const channel = await VoiceOnDemandTools.getVoiceChannel(message.member);
 
     if (!channel) {
       await Tools.handleUserError(
@@ -103,7 +95,7 @@ class HandleUpLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
     }
 
     const limit = Math.min(maxLimit, channel.userLimit + 1);
-    await Tools.updateLimit(channel, limit);
+    await VoiceOnDemandTools.updateLimit(channel, limit);
     await message.reply(
       `Successfully changed the limit of your room to ${limit}`
     );
@@ -122,7 +114,7 @@ class HandleDownLimitCommand implements CommandHandler<DiscordEvent.MESSAGE> {
   async handle(message: Message): Promise<void> {
     const channel = message.member.voice.channel;
     const limit = Math.max(2, channel.userLimit - 1);
-    await Tools.updateLimit(channel, limit);
+    await VoiceOnDemandTools.updateLimit(channel, limit);
     await message.reply(
       `Successfully changed the limit of your room to ${limit}`
     );
